Revisión y corrección general del backend

+++ backend/src/application/use_cases/PricesUseCases/DeletePriceUseCase.ts
@@ -1,4 +1,3 @@
-import { Prisma } from "@prisma/client";
 import { PriceRepository } from "../../../infrastructure/database/repository/PriceRepository.js";
 import { NotFoundError } from "../../../shared/exceptions/NotFoundError.js";
 
@@ -8,13 +7,9 @@ export class DeletePriceUseCase {
     ) {}
 
     public async execute(productId: number, dateFrom: Date): Promise<void> {
-        try {
-            await this.priceRepository.delete(productId, dateFrom);
-        }
-        catch (error) {
-            if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025')
-                throw new NotFoundError("Precio no encontrado");
-            throw error;
-        }
+        const priceFound = await this.priceRepository.getOne(productId, dateFrom);
+        if (!priceFound) throw new NotFoundError("Precio no encontrado");
+
+        await this.priceRepository.delete(productId, dateFrom);
     }
 }
