+++ backend/src/application/use_cases/TableUseCase/DeleteTable.ts
@@ -1,6 +1,5 @@
 import { TableRepository } from "../../../infrastructure/database/repository/TableRepository.js";
 import { NotFoundError } from "../../../shared/exceptions/NotFoundError.js";
-import { Prisma } from "@prisma/client";
 
 export class DeleteTable {
     constructor (
@@ -8,12 +7,10 @@ export class DeleteTable {
     ) {}
 
     public async execute (numTable : number): Promise<void> {
-        try {
-            await this.tableRepository.deleteTable(numTable);
-        } catch (error){
-            if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025')
-                throw new NotFoundError("Mesa no encontrada");
-            throw error;
-        }
+        const tableFound = await this.tableRepository.getByNumTable(numTable);
+
+        if (!tableFound) throw new NotFoundError("Mesa no encontrada");
+
+        await this.tableRepository.deleteTable(numTable);
     }
 }
