+++ backend/src/application/use_cases/WaiterUseCases/CUU24_deleteWaiter.ts
@@ -1,4 +1,3 @@
-import { Prisma } from '@prisma/client';
 import { WaiterRepository } from '../../../infrastructure/database/repository/WaiterRepository.js';
 import { NotFoundError } from '../../../shared/exceptions/NotFoundError.js';
 
@@ -8,13 +7,9 @@ export class CUU24DeleteWaiter {
     ) {}
 
     public async execute(idMozo: string): Promise<void> {
-        try {
-            await this.waiterRepository.deleteWaiter(idMozo);
-        }
-        catch(error) {
-            if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025')
-                throw new NotFoundError("Mozo no encontrado");
-            throw error;
-        }
+        const waiterFound = await this.waiterRepository.getWaiterById(idMozo);
+        if (!waiterFound) throw new NotFoundError("Mozo no encontrado");
+
+        await this.waiterRepository.deleteWaiter(idMozo);
     }
 }
